# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Talk",
    "description": "Talk Project API"
  },
  "host": "localhost:3000",
  "basePath": "/v1/api",
  "tags": [
    {
      "name": "Auth",
      "description": "API for authentication"
    },
    {
      "name": "User",
      "description": "API for get user information"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "signup user",
        "description": "create new user",
        "parameters": [
          {
            "name": "user information",
            "in": "body",
            "description": "user that we want to create",
            "schema": {
              "$ref": "#/definitions/UserSignUp"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "create",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "status": 200,
                "message": ".ثبت نام با موفقیت انجام شد"
              }
            }
          },
          "400": {
            "description": "body validation failed",
            "schema": {
              "type": "array",
              "$ref": "#/definitions/Validation"
            }
          },
          "409": {
            "description": "user already exists",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "status": 409,
                "message": ".کاربر از قبل در سامانه ثبت شده است"
              }
            }
          },
          "429": {
            "description": "rate limiter error",
            "schema": {
              "$ref": "#/components/schemas/rateLimiter"
            }
          },
          "500": {
            "description": "server error",
            "schema": {
              "$ref": "#/components/schemas/serverError"
            }
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "tags": ["Auth"],
        "summary": "sigin user",
        "description": "login user",
        "parameters": [
          {
            "name": "email & password",
            "in": "body",
            "description": "login user to website",
            "schema": {
              "$ref": "#/definitions/UserSignIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "login success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                },
                "token": {
                  "type": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "status": 200,
                "message": "ورود با موفقیت انجام شد",
                "token": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImFkN2FmOTlmLTg1MWYtNDA5OC1iOGE3LTdhZDYyMGNjMzRkNiIsImVtYWlsIjoiemhtdGltMUBnbWFpbC5jb20iLCJwaG9uZSI6IjA5MjEyMzIwNDk5IiwiZnVsbG5hbWUiOiJtb2VpbiIsImlhdCI6MTY1OTM4MDI1NCwiZXhwIjoxNjU5MzgwMjY5fQ.Ccgj_1xrduLHYuFIgQ24KRPiTxznpF8LvQGnE2ytk7Ijv0ixZainppwKjCUR6k2dIJqD_gBbBM1CwGxMQq9hr9VlvXxdet7SsLPqjOLf3h2QxmHlTH9zSQHtN5qRMGud_zhY6rT0T_OlZqSintoTupvMIJS7T37OnH6Xqi-YIfo"
              }
            }
          },
          "400": {
            "description": "body validation failed",
            "schema": {
              "type": "array",
              "$ref": "#/definitions/Validation"
            }
          },
          "403": {
            "description": "login success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "status": 403,
                "message": ".رمز عبور صحیح نمی باشد"
              }
            }
          },
          "404": {
            "description": "login success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "status": 404,
                "message": ".کاربری با این ایمیل یافت نشد"
              }
            }
          },
          "429": {
            "description": "rate limiter error",
            "schema": {
              "$ref": "#/components/schemas/rateLimiter"
            }
          },
          "500": {
            "description": "server error",
            "schema": {
              "$ref": "#/components/schemas/serverError"
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["User"],
        "summary": "get all user",
        "description": "information of all users",
        "responses": {
          "200": {
            "description": "find users",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": "200"
                },
                "message": {
                  "type": "string",
                  "example": "لیست کاربران"
                },
                "users": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/UserInformation"
                  }
                }
              }
            }
          },
          "404": {
            "description": "not found user",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                },
                "users": {
                  "type": "array"
                }
              },
              "example": {
                "status": 404,
                "message": "کاربری یافت نشد",
                "users": []
              }
            }
          },
          "500": {
            "description": "server error",
            "schema": {
              "$ref": "#/components/schemas/serverError"
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": ["User"],
        "summary": "get one user by id",
        "description": "get information user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "type": "string",
            "format": "uuid",
            "example": "52e5e7c6-1908-4091-92c7-ea6be4196d1e",
            "required": true,
            "description": "unique id"
          }
        ],
        "responses": {
          "200": {
            "description": "find user",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": "200"
                },
                "message": {
                  "type": "string",
                  "example": "کاربر مورد نظر یافت شد"
                },
                "user": {
                  "$ref": "#/definitions/UserInformation"
                }
              }
            }
          },
          "400": {
            "description": "params validation failed",
            "schema": {
              "type": "array",
              "$ref": "#/definitions/Validation"
            }
          },
          "404": {
            "description": "not found user",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "status": 404,
                "message": "کاربر مورد نظر یافت نشد"
              }
            }
          },
          "500": {
            "description": "server error",
            "schema": {
              "$ref": "#/components/schemas/serverError"
            }
          }
        }
      }
    }
  },

  "definitions": {
    "Validation": {
      "properties": {
        "status": {
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "field": {
                "type": "string"
              }
            }
          }
        }
      },
      "example": {
        "status": 400,
        "message": "مشکل در تایید اعتبار",
        "data": [
          {
            "type": "required",
            "message": ".فیلد الزامی می باشد",
            "field": "field"
          }
        ]
      }
    },
    "UserSignUp": {
      "required": ["fullname", "email", "phone", "password", "confirmPassword"],
      "properties": {
        "fullname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      },
      "example": {
        "fullname": "moein zhmti",
        "email": "example@gmail.com",
        "phone": "09133216598",
        "password": "abcd12345678",
        "confirmPassword": "abcd12345678"
      }
    },
    "UserSignIn": {
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "example": {
        "email": "example@gmail.com",
        "password": "abcd12345678"
      }
    },
    "UserInformation": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "fullname": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      },
      "example": {
        "id": "636c4ce3-2de2-4e6c-a325-227a725c4132",
        "fullname": "fName lName",
        "email": "example@gmail.com",
        "phone": "09133216598",
        "role": "member"
      }
    }
  },

  "components": {
    "schemas": {
      "serverError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "status": 500,
          "message": ".خطایی در سمت سرور رخ داده است"
        }
      },
      "rateLimiter": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "status": 429,
          "message": "درخواست بیش از حد مجاز"
        }
      }
    }
  }
}